Add goal tree consistency checking.
Because making these can be tedious, I'm adding strict goal
tree checking code that will throw warnings when items are
added to the goal tree that are invalid or incomplete. As
a result I noticed that farming is a dependency already, and
so we add that as well to assure proper operation.